requirements.txt
discord.py==2.4.0
python-dotenv==1.0.1
cryptography==43.0.1
openpyxl==3.1.5
filelock==3.16.0
requests==2.32.3
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0